Fix key error and import from path
