+++ input/demos/boreas.toml
@@ -1,0 +1,170 @@
+# PROTEUS configuration file (version 2.0)
+
+# Metadata
+version = "2.0"
+
+# ----------------------------------------------------
+# Parameters
+[params]
+    [params.out]
+        logging     = "INFO"
+        path        = "boreas"
+        plot_fmt    = "pdf"  # Plotting image file format, "png" or "pdf" recommended
+
+    # time-stepping
+    [params.dt]
+        minimum      = 3e2    # yr, minimum time-step
+        maximum      = 8e7    # yr, maximum time-step
+        initial      = 1e4    # yr, inital step size
+        starspec     = 1e9    # yr, interval to re-calculate the stellar spectrum
+        starinst     = 100    # yr, interval to re-calculate the instellation
+
+        method = 'maximum'
+
+        [params.dt.adaptive]
+            atol         = 0.02   # Step size atol
+            rtol         = 0.07   # Step size rtol
+
+    [params.stop]
+
+        [params.stop.time]
+            enabled = true
+            minimum = 8e8     # yr, model will certainly run to t > minimum
+            maximum = 2e9     # yr, model will terminate when t > maximum
+
+        [params.stop.solid]
+            enabled  = false
+            phi_crit = 0.005  # non-dim., model will terminate when global melt fraction < phi_crit
+
+        [params.stop.radeqm]
+            enabled = false
+            atol    = 0.2     # absolute tolerance [W m-2]
+            rtol    = 1e-3    # relative tolerance
+
+        [params.stop.escape]
+            enabled   = true
+            p_stop    = 1.0  # bar, model will terminate with p_surf < p_stop
+
+
+[star]
+
+    # Physical parameters
+    mass    = 1.0       # M_sun
+    age_ini = 0.100     # Gyr, model initialisation/start age
+
+    module  = "mors"
+    [star.mors]
+        rot_pcntle = 50.0       # rotation percentile
+        tracks  = "spada"   # evolution tracks: spada | baraffe
+        age_now = 4.567     # Gyr, current age of star used for scaling
+        spec    = "stellar_spectra/Named/sun.txt" # stellar spectrum
+
+[orbit]
+    semimajoraxis   = 0.5       # AU
+    eccentricity    = 0.0       # dimensionless
+    zenith_angle    = 48.19     # degrees
+    s0_factor       = 0.375     # dimensionless
+    module  = "none"
+
+[struct]
+    mass_tot    = 1.0           # Total planet mass [M_earth]
+    corefrac    = 0.55          # non-dim., radius fraction
+
+# Atmosphere - physics table
+[atmos_clim]
+    prevent_warming = true      # do not allow the planet to heat up
+    surface_d       = 0.01      # m, conductive skin thickness
+    surface_k       = 2.0       # W m-1 K-1, conductive skin thermal conductivity
+    surf_state      = "fixed"    # surface scheme: "mixed_layer" | "fixed" | "skin"
+    surf_greyalbedo  = 0.1      # surface grey albedo
+    rayleigh        = false      # enable rayleigh scattering
+
+    module  = "agni"           # Which atmosphere module to use
+
+    [atmos_clim.dummy]
+        gamma = 0.98
+
+    [atmos_clim.agni]
+        solve_energy = false
+        spectral_group = "Dayspring"
+        spectral_bands = "16"
+        p_top = 1e-05
+        surf_material = "surface_albedos/Hammond24/lunarmarebasalt.dat"
+        num_levels = 32
+        chemistry = "none"
+        solution_atol = 2.0
+        solution_rtol = 0.13
+        overlap_method = "ee"
+        condensation = false
+        real_gas = true
+
+[escape]
+
+    reservoir = "pxuv"  # Escaping reservoir: "bulk", "outgas", "pxuv".
+    module = "boreas"    # Which escape module to use
+
+    [escape.boreas]
+        light_major = 'H'
+        heavy_major = 'O'
+        heavy_minor = 'CN'
+        efficiency  = 0.1
+        alpha_rec   = 2.6e-13
+        sigma_XUV   = 1.89e-18
+        kappa_p     = 4e-3
+
+    [escape.zephyrus]
+        Pxuv        = 1e-4          # Pressure at which XUV radiation become opaque in the planetary atmosphere [bar]
+        efficiency  = 1.0           # Escape efficiency factor
+        tidal       = false         # Tidal contribution enabled
+
+    [escape.dummy]
+        rate = 1e6
+
+
+[interior]
+    radiogenic_heat = false  # enable radiogenic heat production
+    tidal_heat      = false  # enable tidal heat production
+
+    module = "dummy"   # Which interior module to use
+
+    [interior.dummy]
+        ini_tmagma = 3000.0
+
+[outgas]
+    fO2_shift_IW    = 5         # log10(ΔIW), atmosphere/interior boundary oxidation state
+
+    module = "calliope"         # Which outgassing module to use
+
+    [outgas.calliope]
+        include_H2O  = true     # Include H2O compound
+        include_CO2  = true     # Include CO2 compound
+        include_N2   = true     # Include N2 compound
+        include_S2   = true     # Include S2 compound
+        include_SO2  = true    # Include SO2 compound
+        include_H2S  = true     # Include H2S compound
+        include_NH3  = true     # Include NH3 compound
+        include_H2   = true    # Include H2 compound
+        include_CH4  = true    # Include CH4 compound
+        include_CO   = true    # Include CO compound
+
+[delivery]
+
+    # Which initial inventory to use?
+    initial = 'elements'        # "elements" | "volatiles"
+
+    # No module for accretion as of yet
+    module = "none"
+
+    # Set initial volatile inventory by partial pressures in atmosphere
+    [delivery.elements]
+        H_oceans = 30.0
+        CH_ratio = 1.0
+        SH_ratio = 2.0
+        N_ppmw   = 3.0
+
+# Calculate simulated observations
+[observe]
+    synthesis = "none"
+
+[atmos_chem]
+    module = "none"
